Motivation, don't leave me now

+++ Assets/Scripts/WeatherSystem/ResultModels/Entities/Wind.cs
@@ -5,6 +5,6 @@ namespace WeatherSystem.ResultModels.Entities
     public class Wind
     {
         [JsonProperty("speed")] public double Speed { get; set; }
-        [JsonProperty("deg")] public int Degrees { get; set; }
+        [JsonProperty("deg")] public double Degrees { get; set; }
     }
 }
